Stub out the RSS project (#1)
* Stub out the RSS project

A number of stub files were added for classes that are expected to be needed in the RSS project.  A few additional files were copied from the arxiv-zero project and modified for use here.  Sample blueprint code was copied from the arxiv-browse project in order to confirm that the development environment and initial routes code was being executed and running correctly.

Later submissions will build the RSS functionality from the top down.

* Respond to review comments

Respond to reviewer comments and conversations with the following changes:
* Use of celery was removed from the project.
* Documentation and whitespace were modified.
* Multiple routes were simplified to just one, without a trailing "/".
* The mypy configuration was updated to a more modern version.
* Added a test for types, copied from arxiv-browse.