Arxivng 1138 - Initial implementation for RSS 2.0 serialization (#2)
* Stub out the RSS project

A number of stub files were added for classes that are expected to be needed in the RSS project.  A few additional files were copied from the arxiv-zero project and modified for use here.  Sample blueprint code was copied from the arxiv-browse project in order to confirm that the development environment and initial routes code was being executed and running correctly.

Later submissions will build the RSS functionality from the top down.

* Respond to review comments

Respond to reviewer comments and conversations with the following changes:
* Use of celery was removed from the project.
* Documentation and whitespace were modified.
* Multiple routes were simplified to just one, without a trailing "/".
* The mypy configuration was updated to a more modern version.
* Added a test for types, copied from arxiv-browse.

* Stub out the RSS project (#1)

* Stub out the RSS project

A number of stub files were added for classes that are expected to be needed in the RSS project.  A few additional files were copied from the arxiv-zero project and modified for use here.  Sample blueprint code was copied from the arxiv-browse project in order to confirm that the development environment and initial routes code was being executed and running correctly.

Later submissions will build the RSS functionality from the top down.

* Respond to review comments

Respond to reviewer comments and conversations with the following changes:
* Use of celery was removed from the project.
* Documentation and whitespace were modified.
* Multiple routes were simplified to just one, without a trailing "/".
* The mypy configuration was updated to a more modern version.
* Added a test for types, copied from arxiv-browse.

* Initial implementation of RSS serialization

This submission includes several changes that were recommended by Erick, plus an initial implementation of the RSS 2.0 serialization.  The indexing is merely a "match_all" operation.  Two files were renamed, though it looks like they may be treated as add/deletes.

* Remove some files whose names were changed

* Respond to PR review comments

These changes are in response to reviewer comments on this pull request, plus some housekeeping.  They include:
* Make it easier for the code to be run in PyCharm to simplify debugging.
* Update the paper IDs that are used to gather suitable test data (these papers include all features that are exported to RSS and Atom.
* Pin package versions in the pipfile.
* Use url_for (with appropriate configuration) to generate arXiv URLs.
* Move some string constants into variables.
* Remove two unwanted elements from the RSS output.
* Use f.strings to format strings from data.
* Add author links to RSS output.
* Convert current time to UTC.