Initial implementation of Atom serialization (#3)
* Initial implementation of Atom serialization

This is an initial implementation of Atom serialization using the feedgen package.  It is nearly a complete reproduction of the existing feed's serialization, with the following difference:
* It does not include affiliation elements for the authors.
* It does include a generator element at the top level (created automatically by feedgen).
* The arxiv namespace is defined in the top element only, not within each arxiv element of the output.

This implementation may have some of the same flaws that were found in the initial RSS implementation, and these will be fixed with upcoming submissions.

The atom_extension class allows the feedgen package to include arxiv specific elements in an item.  Unfortunately, this package does not provide a way to add custom elements to the authors, which would be needed to include the affiliation element.  We need to determine whether we want to extend feedgen to do this or rewrite this code to use another package.

* Arxivng 1886: adding travis-ci config (#4)

* ARXIVNG-1886: copying over arxiv-browse .travis.yml

* ARXIVNG-1886: copying over travis scripts and adding dev dependencies

* ARXIVNG-1886: updating Pipfile.lock to hopefully address elasticsearch failure

* Code cleanup to meet standards

Modify existing code so that it passess muster with mypy, pylint and pydocstyle.  This included reording imports, breaking long lines of code, adjusting parameter types and adding lots of docstyle headers and content.  Imports were all put into "from X import Y" form.

The MyPyTest test fails on my machine, but I'd like to see how it works for Travis.  Will attempt to get this working before mergin my PR.

* making test shell scripts executable

* reverting back to testing on rss