Merge work to date in Develop into Master (#6)
* Stub out the RSS project (#1)

* Stub out the RSS project

A number of stub files were added for classes that are expected to be needed in the RSS project.  A few additional files were copied from the arxiv-zero project and modified for use here.  Sample blueprint code was copied from the arxiv-browse project in order to confirm that the development environment and initial routes code was being executed and running correctly.

Later submissions will build the RSS functionality from the top down.

* Respond to review comments

Respond to reviewer comments and conversations with the following changes:
* Use of celery was removed from the project.
* Documentation and whitespace were modified.
* Multiple routes were simplified to just one, without a trailing "/".
* The mypy configuration was updated to a more modern version.
* Added a test for types, copied from arxiv-browse.

* Initial implementation of Atom serialization (#3)

* Initial implementation of Atom serialization

This is an initial implementation of Atom serialization using the feedgen package.  It is nearly a complete reproduction of the existing feed's serialization, with the following difference:
* It does not include affiliation elements for the authors.
* It does include a generator element at the top level (created automatically by feedgen).
* The arxiv namespace is defined in the top element only, not within each arxiv element of the output.

This implementation may have some of the same flaws that were found in the initial RSS implementation, and these will be fixed with upcoming submissions.

The atom_extension class allows the feedgen package to include arxiv specific elements in an item.  Unfortunately, this package does not provide a way to add custom elements to the authors, which would be needed to include the affiliation element.  We need to determine whether we want to extend feedgen to do this or rewrite this code to use another package.

* Arxivng 1886: adding travis-ci config (#4)

* ARXIVNG-1886: copying over arxiv-browse .travis.yml

* ARXIVNG-1886: copying over travis scripts and adding dev dependencies

* ARXIVNG-1886: updating Pipfile.lock to hopefully address elasticsearch failure

* Code cleanup to meet standards

Modify existing code so that it passess muster with mypy, pylint and pydocstyle.  This included reording imports, breaking long lines of code, adjusting parameter types and adding lots of docstyle headers and content.  Imports were all put into "from X import Y" form.

The MyPyTest test fails on my machine, but I'd like to see how it works for Travis.  Will attempt to get this working before mergin my PR.

* making test shell scripts executable

* reverting back to testing on rss