The RSS feed's indexer now returns only eprints that belong to the identified set of archives/categories, and which were added within the last day.  This addresses both issues 1676 and 1677.
The indexer was substantially upgraded to validate the list of requested archives/categories, create an Elasticsearch query that selects for those categories and filter the results based on the proper date range.  A set of archives/categories may be specified by concatenating them with the delimiter '+'.

Error messages are returned as needed and are suitable for display in the browser.

The serializers no longer depend on a non-empty set of eprints when determining the title of the Atom or RSS output.