The Atom serialization did not provide a way to include author affiliations, which are defined in the arXiv schema.  The FeedGenerator extension for Atom has been augmented to allow this information to be collected and serialized for each author.
A list of affiliations was added to the domain model for an author and is populated by the indexer.  The serializer passes this information on to the extension, which prints the affiliation elements during serialization.

The automatically included "generator" element is also removed from the output, and several method signatures and docstrings were corrected.